improved error message for self signed SSL
Problem:
When cyberpanel fails to get a valid SSL certificate from Let's Encrypt, it will generate a self signed certificate.
An error message is logged to the error log, but a normal success message is sent to the user and displayed in the web UI.

The user should be informed right away that the cert generated was self signed, since chrome and firefox will give warning in the browser for most self signed certificates.

Solution:
Give the user a message in the control panel that will indicate that a self signed cert was generated. This way they can investigate further if the desire.