Update get_cyberpanel_version to handle non-JSON data
Previously, `get_cyberpanel_version` could only interpret JSON formatted data from `version.txt`. It would fail if the data was provided as plain text lines.

This commit updates the function to first attempt to interpret the data as JSON. If that fails, it will then try to interpret the data as plain text lines.

This allows for greater flexibility in the format of the data in `version.txt`, making the function more robust to changes in the way version information is stored.